+++ components/Chat/Chat.tsx
@@ -18,6 +18,7 @@ import {
   saveConversations,
   updateConversation,
 } from '@/utils/app/conversation';
+import { showAPIToastError } from '@/utils/app/errors';
 import { throttle } from '@/utils/data/throttle';
 
 import { OpenAIModel } from '../../types/openai';
@@ -34,6 +35,8 @@ import { MemoizedChatMessage } from './MemoizedChatMessage';
 import { ModelSelect } from './ModelSelect';
 import { SystemPrompt } from './SystemPrompt';
 import { TemperatureSlider } from './Temperature';
+
+import { errorsMessages } from '@/constants/errors';
 
 interface Props {
   stopConversationRef: MutableRefObject<boolean>;
@@ -162,7 +165,8 @@ export const Chat = memo(({ stopConversationRef }: Props) => {
         if (!response.ok) {
           homeDispatch({ field: 'loading', value: false });
           homeDispatch({ field: 'messageIsStreaming', value: false });
-          toast.error(response.statusText);
+
+          await showAPIToastError(response, t(errorsMessages.generalServer));
           return;
         }
         const data = response.body;