+++ components/Chat/ChatInput.tsx
@@ -26,6 +26,8 @@ import HomeContext from '@/pages/api/home/home.context';
 
 import { PluginSelect } from './PluginSelect';
 import { PromptList } from './PromptList';
+import { ReportIssueDialog } from './ReportIssueDialog';
+import { RequestAPIKeyDialog } from './RequestApiKeyDialog';
 import { VariableModal } from './VariableModal';
 
 interface Props {
@@ -61,6 +63,8 @@ export const ChatInput = ({
   const [variables, setVariables] = useState<string[]>([]);
   const [isModalVisible, setIsModalVisible] = useState(false);
   const [showPluginSelect, setShowPluginSelect] = useState(false);
+  const [isRequestAPIDialogOpen, setIsRequestAPIDialogOpen] = useState(false);
+  const [isReportIssueDialogOpen, setIsReportIssueDialogOpen] = useState(false);
   const [plugin, setPlugin] = useState<Plugin | null>(null);
 
   const promptListRef = useRef<HTMLUListElement | null>(null);
@@ -392,12 +396,59 @@ export const ChatInput = ({
         {t(
           "Chatbot UI is an advanced chatbot kit for OpenAI's chat models aiming to mimic ChatGPT's interface and functionality.",
         )} */}
-        <a href="https://kb.epam.com/display/EPMGPT/EPAM+AI+Chat" target="_blank" rel="noreferrer"
-          className="underline font-bold">EPAM AI Chat</a> can be used <span className='underline'>any work-related activity</span>. Rest assured,
-        information you share here is{' '}
-        <span className='underline'>not disclosed to third-party companies</span>. However, we{' '}
-        <span className='underline'>anonymize and log</span> all interactions for research purposes. <br />
+        <a
+          href="https://kb.epam.com/display/EPMGPT/EPAM+AI+Chat"
+          target="_blank"
+          rel="noreferrer"
+          className="underline font-bold"
+        >
+          EPAM AI Chat
+        </a>{' '}
+        can be used <span className="underline">any work-related activity</span>
+        . Rest assured, information you share here is{' '}
+        <span className="underline">
+          not disclosed to third-party companies
+        </span>
+        . However, we <span className="underline">anonymize and log</span> all
+        interactions for research purposes. <br />
+        For API access please fill&nbsp;
+        <a
+          href=""
+          onClick={(e) => {
+            e.preventDefault();
+            setIsRequestAPIDialogOpen(true);
+          }}
+          className="underline font-bold"
+        >
+          {t('this form')}
+        </a>
+        .&nbsp;If you have a problem please&nbsp;
+        <a
+          href=""
+          onClick={(e) => {
+            e.preventDefault();
+            setIsReportIssueDialogOpen(true);
+          }}
+          className="underline font-bold"
+        >
+          {t('report an issue')}
+        </a>
+        .
       </div>
+
+      <RequestAPIKeyDialog
+        isOpen={isRequestAPIDialogOpen}
+        onClose={() => {
+          setIsRequestAPIDialogOpen(false);
+        }}
+      ></RequestAPIKeyDialog>
+
+      <ReportIssueDialog
+        isOpen={isReportIssueDialogOpen}
+        onClose={() => {
+          setIsReportIssueDialogOpen(false);
+        }}
+      ></ReportIssueDialog>
     </div>
   );
 };